fix(engine): handle failing build script (#188)
* Handle failing build script

When the build script fails for some reason, the EngineNode hung, because the
:EXIT would not be received, as the Port was started in some other process, but
not in the EngineNode's GenServer. Only the latter explicitely traps exits.

To fix this without causing any side effects (handle_info(:EXIT) in Node), we
run the Port in a Task subprocess that traps exits.

Additionally, we log the last line received from the build script to help the
user to debug the issue, instead of failing silently with exit reason :normal.

* fix: add debug logs for project engine build

* refactor: start pipe with raw value

---------

Co-authored-by: doorgan <dorgandash@gmail.com>